Fix infinite loop: don't cache intercept HTML
If a user visits a protected page without an Anubis cookie, we inject
intercept HTML to send them to Anubis first; once they get their cookie,
Anubis redirects them back to that same protected page.

Previously, we allowed the browser to cache the intercept on that page.
This meant that when Anubis redirected the user back to the protected
page, their browser might load the intercept from cache and send them
back to Anubis in an infinite loop, even though they have a valid cookie
now. This commit breaks the loop by sending anti-cache headers right
before the intercept HTML.